move schema for sign in to login form component and write test for logging process
