+++ frontend/src/components/LoginForm/LoginForm.test.tsx
@@ -1,0 +1,14 @@
+// import { Provider } from "react-redux"
+
+// import { store } from "@/store/store"
+// import { BrowserRouter } from "react-router-dom"
+// import { LoginForm } from "./LoginForm"
+// const MockLoginForm = () => {
+//     return (
+//         <Provider store={store}>
+//             <BrowserRouter>
+//                 <LoginForm/>
+//             </BrowserRouter>
+//         </Provider>
+//     )
+// }
