+++ frontend/src/components/RegistrationForm/RegistrationForm.types.ts
@@ -16,15 +16,3 @@ export const signUpSchema = z
     });
 
 export type TSignUpSchema = z.infer<typeof signUpSchema>;
-
-export const signInSchema = z
-    .object({
-        email: z.string().email('Niepoprawny format adresu e-mail'),
-        password: z.string().nonempty({ message: 'Hasło jest wymagane' }),
-    })
-    .refine((data) => data.email !== '', {
-        message: 'Adres e-mail jest wymagany',
-        path: ['email'],
-    });
-
-export type TSignInSchema = z.infer<typeof signInSchema>;
