+++ frontend/src/routes/Login/Login.test.tsx
@@ -5,7 +5,8 @@ import { Provider } from 'react-redux';
 
 import App from '@/App';
 
-import { store } from '@store/store';
+import { store } from '../../store/store';
+import userEvent from '@testing-library/user-event';
 
 const MockLoginPanel = () => {
     return (
@@ -15,6 +16,12 @@ const MockLoginPanel = () => {
             </MemoryRouter>
         </Provider>
     );
+};
+
+const CORRECT_USER = {
+    name: 'Adam',
+    email: 'xyz@gmail.com',
+    password: 'P@ssw0rd1!',
 };
 
 describe('Test for Login panel', () => {
@@ -40,5 +47,20 @@ describe('Test for Login panel', () => {
         expect(
             await screen.findByRole('heading', { level: 2 })
         ).toHaveTextContent('Stwórz swoje konto już dzisiaj');
+    });
+    it('Should move to dashboard view after proper credentials without any error messages below login form inputs', async () => {
+        render(<MockLoginPanel />);
+        const email = screen.getByPlaceholderText('adam.kowalski@gmail.com');
+        const password = screen.getByPlaceholderText('********');
+        const submit = screen.getByRole('button');
+
+        userEvent.type(email, CORRECT_USER.email);
+        userEvent.type(password, CORRECT_USER.password);
+        fireEvent.click(submit);
+
+        // expect(await screen.findByText('Hasło jest wymagane')).toBeFalsy();
+        expect(
+            await screen.findByRole('heading', { level: 1 })
+        ).toHaveTextContent(`Witaj ${CORRECT_USER.name}`);
     });
 });
